requirements.txt
flask-restplus>=0.9.2,<1.0.0
Flask-SQLAlchemy>=2.1,<3.0
psycopg2>=2.6.2,<3.0.0
flask-cors>=3.0.2,<4.0.0
graphene>=1.1.3,<2.0.0
graphene-sqlalchemy>=1.1.1,<2.0.0
Flask-GraphQL>=1.4.0,<2.0.0
